ImageDialogCanvas: refactor use of makeStyles to use of styled components
ImageDialogCanvas: refactor use of makeStyles to use of styled components
