--- src/components/ImageCanvas/index.ts
@@ -1,1 +1,0 @@
-export { ImageCanvas } from "./ImageCanvas";
