+++ src/components/ImageDialogCanvas/index.ts
@@ -1,1 +1,3 @@
+// TODO: ImageDialogCanvas and associated components (all in this directory) are unused
+
 export { ImageDialogCanvas } from "./ImageDialogCanvas";
